Testing out a thing with Godot
I want to see if Godot can easily retrieve JSON data from a GitHub Pages webpage. I assume it can, which would make for a free way to store information for my games to load in real time. For example, messages that appear at different times, a link to a sequel, or even a link to a video I want to promote to the players. IDK if there's more I would want, but I could make fun little things with this and with reasonably low effort. Plus it would make for a cool tutorial